backend/requirements.txt
fastapi
uvicorn
duckdb
networkx
groq
python-dotenv
pydantic
